Move painters to image folder
